moved util.run_residue to gibbs.run_residue
