+++ basicrta/util.py
@@ -472,19 +472,6 @@ def plot_protein(residues, t_slow, bars, prot=None, label_cutoff=3, ylim=None,
 #     gib.run()
 
 
-def run_residue(residue, time, proc, ncomp, niter, cutoff, g, from_combined=False):
-    from basicrta.gibbs import Gibbs
-    x = np.array(time)
-    if len(x) != 0:
-        try:
-            proc = int(multiprocessing.current_process().name.split('-')[-1])
-        except ValueError:
-            proc = 1
-
-    gib = Gibbs(x, residue, proc, ncomp=ncomp, niter=niter, cutoff=cutoff)
-    gib._from_combined_contacts = from_combined
-    gib.run()
-
 
 def check_results(residues, times, ts):
     if not os.path.exists('result_check'):
